feat: Exception Handling Group

+++ src/storage/group/groupCreate.ts
@@ -1,12 +1,19 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@utils/AppErros";
 
 export async function GroupCreate(newGroup: string) {
   try {
-    const storaedGroup = await groupsGetAll();
+    const storedGroup = await groupsGetAll();
 
-    const storage = JSON.stringify([...storaedGroup, newGroup]);
+    const groupAlreadyExists = storedGroup.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError("Já existe um grupo cadastrado com esse nome");
+    }
+
+    const storage = JSON.stringify([...storedGroup, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
     throw error;
